Moved strategies.
